SubGHz: extract Add Manually into a loadable plugin
Add Manually and Add Manually [Advanced] are one flow of six scenes plus the
generator table behind them, and none of it is touched unless someone is building
a remote by hand - but all of it sat in the firmware image. It now ships as
subghz_add_manually.fal, mapped when the flow is entered and dropped once it is
done, which frees 9,584 bytes of flash: usable free flash goes from 16,748 to
26,332 bytes.

The six scenes stay in the app as shims that forward to the plugin, because the
scene handler tables are built at compile time and save_name still looks for
SubGhzSceneSetType on the stack. That is the same shape the NFC app uses for its
plugin-provided scenes, hand-written thunks included. Everything else - the
96-entry protocol menu, the GenInfo filler and the nineteen generators - moved.

Unmapping is deliberately not symmetric with mapping. The flow's scenes push one
another, so a forward transition runs the current scene's on_exit from inside
plugin code; dropping the image there would carry on executing freed memory. The
start scene is where every exit from the flow arrives with none of its scenes
left on the stack - after a save, by way of the saved list - so it unmaps there,
and app teardown does the same for a kill that runs no scene handler.

Unmapping also has to take back what the plugin left in the app's views:
byte_input keeps its header string by reference and draws through it, and the
string the last scene installed is a literal inside the image - including the
empty one its own on_exit puts there.

GenInfo stays app-allocated, as it was, because byte_input holds a pointer into
it. Its contents do not survive an unmap either - mod, name and manuf point into
the plugin - but nothing reads them across one: re-entering the flow refills them.

Only two symbols could not be resolved from the firmware SDK, so the app's
private table gains subghz_txrx_set_preset and subghz_file_name_clear; the
fourteen subghz_protocol_*_create_data functions the generators call were already
exported. The APPCHK warning naming those two matches the one the Frequency
Analyzer plugin already produces - the import check cannot see a host app's
private table.
