+++ applications/main/subghz/api/subghz_app_api_table_i.h
@@ -13,4 +13,8 @@ static constexpr auto subghz_app_api_table = sort(create_array_t<sym_entry>(
     // levels on the way out.
     API_METHOD(subghz_txrx_get_setting, SubGhzSetting*, (SubGhzTxRx*)),
     API_METHOD(subghz_txrx_radio_device_is_frequency_valid, bool, (SubGhzTxRx*, uint32_t)),
-    API_METHOD(subghz_last_settings_save, bool, (SubGhzLastSettings*))));
+    API_METHOD(subghz_last_settings_save, bool, (SubGhzLastSettings*)),
+    // Add Manually: every generator sets the radio preset before serializing, and each scene that
+    // finishes a remote clears the loaded file name so the save lands on a fresh one.
+    API_METHOD(subghz_txrx_set_preset, void, (SubGhzTxRx*, const char*, uint32_t, uint8_t*, size_t)),
+    API_METHOD(subghz_file_name_clear, void, (SubGhz*))));
