+++ applications/main/subghz/subghz_i.h
@@ -38,6 +38,7 @@
 #include "helpers/subghz_threshold_rssi.h"
 
 #include "helpers/subghz_txrx.h"
+#include "helpers/subghz_add_manually_plugin.h"
 #include "helpers/subghz_frequency_analyzer_plugin.h"
 
 #include <flipper_application/plugins/composite_resolver.h>
@@ -93,12 +94,15 @@ struct SubGhz {
     SubGhzRxKeyState rx_key_state;
     SubGhzHistory* history;
 
-    // Frequency Analyzer, resident only while its scene is current. The resolver is refcounted
-    // rather than owned by whichever plugin loaded first, so a second one can share it.
+    // Feature plugins, mapped on demand: the analyzer while its scene is current, Add Manually
+    // until the start scene is reached again. The resolver they share is refcounted rather than
+    // owned by whichever loaded first.
     CompositeApiResolver* api_resolver;
     uint8_t api_resolver_refs;
     PluginManager* freq_analyzer_plugin_manager;
     const SubGhzFrequencyAnalyzerPlugin* freq_analyzer_plugin;
+    PluginManager* add_manually_plugin_manager;
+    const SubGhzAddManuallyPlugin* add_manually_plugin;
 
     uint16_t idx_menu_chosen;
     SubGhzLoadTypeFile load_type_file;
