+++ applications/main/subghz_add_manually/application.fam
@@ -1,0 +1,8 @@
+App(
+    appid="subghz_add_manually",
+    apptype=FlipperAppType.PLUGIN,
+    targets=["f7"],
+    entry_point="subghz_add_manually_ep",
+    requires=["subghz"],
+    sources=["*.c"],
+)
